+++ src/devices/ht680.c
@@ -18,7 +18,7 @@ static int ht680_callback(r_device *decoder, bitbuffer_t *bitbuffer)
     data_t *data;
     uint8_t b[5]; // 36 bits
 
-    for (uint8_t row = 0;row < bitbuffer->num_rows;row++){
+    for (int row = 0; row < bitbuffer->num_rows; row++) {
         if (bitbuffer->bits_per_row[row] != 41 || // Length of packet is 41 (36+5)
                 (bitbuffer->bb[row][0] & 0xf8) != 0xa8) // Sync is 10101xxx (5 bits)
         continue;
